requirements.txt
Django==4.2.1
django-rest-passwordreset==1.3.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.5
Pillow==9.5.0
rest-framework-simplejwt==0.0.2

